+++ src/tensor/ops.rs
@@ -29,10 +29,10 @@ pub enum OpType {
 }
 
 pub trait TensorConstructors {
-    fn new(a: Array2<f32>) -> Tensor;
+    fn new(a: Array2<f64>) -> Tensor;
     fn ones<Sh>(shape: Sh) -> Tensor where Sh: ShapeBuilder<Dim = Dim<[usize; 2]>>;
     fn zeros<Sh>(shape: Sh) -> Tensor where Sh: ShapeBuilder<Dim = Dim<[usize; 2]>>;
-    fn fill<Sh>(shape: Sh, x: f32) -> Tensor where Sh: ShapeBuilder<Dim = Dim<[usize; 2]>>;
+    fn fill<Sh>(shape: Sh, x: f64) -> Tensor where Sh: ShapeBuilder<Dim = Dim<[usize; 2]>>;
     fn randn<Sh>(shape: Sh) -> Tensor where Sh: ShapeBuilder<Dim = Dim<[usize; 2]>>;
 }
